login and register

+++ app/Http/Controllers/DoctorController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Specialization;
 use App\Repositories\UserRepository;
+use Illuminate\Support\Facades\Auth;
 
 class DoctorController extends Controller
 {
@@ -17,10 +18,18 @@ class DoctorController extends Controller
     public function __construct(UserRepository $userRepo)
     {
         $this->userRepo = $userRepo;
+        //wywoływanie kontrolera dostepu middleware
+        //dodanie filtru auth - bedzie sprawdzal czy user jest zalogowany
+        $this->middleware('auth');
     }
 
     public function index()
     {
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }       
+        
         $statistics = $this->userRepo->getDoctorsStatistics();
 
         $user = $this->userRepo->getAllDoctors();
@@ -36,6 +45,11 @@ class DoctorController extends Controller
 
     public function listBySpecialization($id)
     {
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+        
         $statistics = $this->userRepo->getDoctorsStatistics();
 
         $user = $this->userRepo->getDoctorsBySpecializations($id);
@@ -54,6 +68,11 @@ class DoctorController extends Controller
         //$user = User::all();
 
         //pojedynczy uzytkownik
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+
         $doctor = $this->userRepo->find($id);
 
         return view('doctors.show', [
@@ -65,7 +84,10 @@ class DoctorController extends Controller
 
     public function store(Request $request) 
     {
-        
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
         //walidowanie danych
         //błedy sa dodawane do sesji i mozna wykorzystac je w formualrzu
         //$erros->any(), $errors->all()
@@ -98,7 +120,12 @@ class DoctorController extends Controller
 
     //create doctor from form
     public function create()
-    {        
+    {       
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+
         $specializations = Specialization::all();
 
         return view('doctors.create', ["specializations"=>$specializations, "footerDate" => Date('Y')]);
@@ -121,6 +148,11 @@ class DoctorController extends Controller
     //create doctor from form
     public function edit($id)
     {
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+
         $doctor = $this->userRepo->find($id);
         $specializations = Specialization::all();
         return view('doctors.edit', ["specializations"=>$specializations, "doctor"=>$doctor, "footerDate" => Date('Y')]);
@@ -129,6 +161,11 @@ class DoctorController extends Controller
 
     public function editStore(Request $request) 
     {
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+
         //stworzenie nowego obiektu
         $doctor = User::find($request->input('doctorId'));
         $doctor->name = $request->input('name');
@@ -146,6 +183,11 @@ class DoctorController extends Controller
     
     public function delete($id)
     {
+        //autoryzacja logowania
+        if(Auth::user()->type != 'doctor' && Auth::user()->type != 'admin') {
+            return redirect()->route('login');
+        }    
+        
         $doctor = $this->userRepo->delete($id);
         return redirect('doctors');
     }
